--- tests/e2e/color-picker-persistence.spec.js
@@ -1,283 +1,0 @@
-// tests/e2e/color-picker-persistence.spec.js
-
-import { test, expect } from '@playwright/test';
-import { CanvasPage, TestCoordinates } from './helpers/CanvasPage.js';
-
-test.describe('Color Picker - Data Persistence', () => {
-  let canvasPage;
-
-  test.beforeEach(async ({ page }) => {
-    canvasPage = new CanvasPage(page);
-    await canvasPage.load();
-  });
-
-  test('Should persist colors across page refresh @smoke', async ({ page }) => {
-    // Create notes with different colors
-    const note1 = await canvasPage.createNote(
-      TestCoordinates.note1.x,
-      TestCoordinates.note1.y,
-    );
-    await canvasPage.selectNote(note1);
-    await page.click('.color-swatch[data-color="pink"]');
-    await canvasPage.editNoteContent('Pink note', note1);
-
-    const note2 = await canvasPage.createNote(
-      TestCoordinates.note2.x,
-      TestCoordinates.note2.y,
-    );
-    await canvasPage.selectNote(note2);
-    await page.click('.color-swatch[data-color="blue"]');
-    await canvasPage.editNoteContent('Blue note', note2);
-
-    // Clear selection before setting global color by creating a temporary note
-    const tempNote = await canvasPage.createNote(500, 500);
-    await tempNote.press('Delete'); // Delete the temporary note
-    await page.waitForTimeout(100); // Small wait for cleanup
-
-    // Set current color to green
-    await page.click('.color-swatch[data-color="green"]');
-
-    // Verify colors before refresh
-    await expect(note1).toHaveClass(/color-pink/);
-    await expect(note2).toHaveClass(/color-blue/);
-    await expect(page.locator('.color-swatch[data-color="green"]')).toHaveClass(
-      /active/,
-    );
-
-    // Refresh the page
-    await page.reload();
-    await page.waitForLoadState('domcontentloaded');
-    await canvasPage.waitForAppReady();
-
-    // Find notes by their content after refresh
-    const pinkNote = page.locator('.note:has-text("Pink note")');
-    const blueNote = page.locator('.note:has-text("Blue note")');
-
-    // Verify colors persisted
-    await expect(pinkNote).toHaveClass(/color-pink/);
-    await expect(blueNote).toHaveClass(/color-blue/);
-
-    // Verify current color state persisted
-    await expect(page.locator('.color-swatch[data-color="green"]')).toHaveClass(
-      /active/,
-    );
-  });
-
-  test('Should maintain colors in export/import workflow', async ({ page }) => {
-    // Mock clipboard to avoid browser permission issues
-    await page.evaluate(() => {
-      let clipboardData = '';
-      Object.defineProperty(navigator, 'clipboard', {
-        value: {
-          writeText: (text) => {
-            clipboardData = text;
-            return Promise.resolve();
-          },
-          readText: () => Promise.resolve(clipboardData)
-        },
-        writable: true
-      });
-    });
-
-    // Create notes with specific colors
-    const note1 = await canvasPage.createNote(300, 300);
-    await canvasPage.selectNote(note1);
-    await page.click('.color-swatch[data-color="pink"]');
-    await canvasPage.editNoteContent('Exported pink', note1);
-
-    const note2 = await canvasPage.createNote(500, 300);
-    await canvasPage.selectNote(note2);
-    await page.click('.color-swatch[data-color="blue"]');
-    await canvasPage.editNoteContent('Exported blue', note2);
-
-    // Connect the notes to make export more meaningful
-    await canvasPage.connectNotes(note1, note2);
-
-    // Export to clipboard
-    await page.click('#kebab-menu-button');
-    await page.click('.kebab-menu-item[data-action="copy-clipboard"]');
-
-    // Wait for export notification
-    await expect(page.locator('.notification-toast')).toBeVisible();
-    await page.waitForTimeout(1000);
-
-    // Clear canvas
-    await page.click('#kebab-menu-button');
-    await page.click('.kebab-menu-item[data-action="clear-canvas"]');
-    await page.click('.notification-modal-button-confirm');
-
-    // Wait for canvas to be cleared
-    await canvasPage.waitForCanvasEmpty();
-
-    // Import from clipboard
-    await page.click('#kebab-menu-button');
-    await page.click('.kebab-menu-item[data-action="paste-clipboard"]');
-
-    // Wait for import to complete
-    await page.waitForTimeout(1500);
-
-    // Find imported notes by content
-    const importedPinkNote = page.locator('.note:has-text("Exported pink")');
-    const importedBlueNote = page.locator('.note:has-text("Exported blue")');
-
-    // Verify colors were restored
-    await expect(importedPinkNote).toBeVisible();
-    await expect(importedPinkNote).toHaveClass(/color-pink/);
-    await expect(importedBlueNote).toBeVisible();
-    await expect(importedBlueNote).toHaveClass(/color-blue/);
-
-    // Verify connection was also restored
-    await expect(page.locator('g[data-start][data-end]')).toBeVisible();
-  });
-
-  test.skip('Should handle backwards compatibility with files without color data', async ({
-    page,
-  }) => {
-    // Simulate importing a file without color information
-    // (This would normally come from a file, but we'll create it programmatically)
-
-    // First create some content with colors
-    const note1 = await canvasPage.createNote(400, 300);
-    await canvasPage.selectNote(note1);
-    await page.click('.color-swatch[data-color="green"]');
-    await canvasPage.editNoteContent('Green note', note1);
-
-    // Clear the canvas
-    await page.click('#kebab-menu-button');
-    await page.click('.kebab-menu-item[data-action="clear-canvas"]');
-    await page.click('.notification-modal-button-confirm');
-    await canvasPage.waitForCanvasEmpty();
-
-    // Instead of importing, manually create a note without color to test backwards compatibility
-    // This simulates what would happen with legacy data
-    await page.evaluate(() => {
-      // Manually create a legacy note by manipulating the DOM
-      // This simulates the result of importing old format data
-      const canvas = document.getElementById('canvas');
-      const note = document.createElement('div');
-      note.className = 'note color-yellow'; // Default color for legacy notes
-      note.id = 'note-1';
-      note.style.left = '400px';
-      note.style.top = '300px';
-      note.style.width = '120px';
-      note.style.height = '80px';
-
-      const content = document.createElement('div');
-      content.className = 'note-content';
-      content.textContent = 'Legacy note';
-      note.appendChild(content);
-
-      canvas.appendChild(note);
-    });
-
-    await page.waitForTimeout(1000);
-
-    // Find the imported note
-    const legacyNote = page.locator('.note:has-text("Legacy note")');
-    await expect(legacyNote).toBeVisible();
-
-    // Should have default yellow color (backwards compatibility)
-    await expect(legacyNote).toHaveClass(/color-yellow/);
-
-    // Color picker should still work with legacy note
-    await canvasPage.selectNote(legacyNote);
-    await page.click('.color-swatch[data-color="blue"]');
-    await expect(legacyNote).toHaveClass(/color-blue/);
-  });
-
-  test('Should preserve colors when switching templates', async ({ page }) => {
-    // Create notes with colors on Standard Canvas
-    const note1 = await canvasPage.createNote(400, 300);
-    await canvasPage.selectNote(note1);
-    await page.click('.color-swatch[data-color="pink"]');
-    await canvasPage.editNoteContent('Pink on standard', note1);
-
-    const note2 = await canvasPage.createNote(600, 300);
-    await canvasPage.selectNote(note2);
-    await page.click('.color-swatch[data-color="green"]');
-    await canvasPage.editNoteContent('Green on standard', note2);
-
-    // Switch to Hero's Journey template
-    await canvasPage.switchToTemplate("Hero's Journey");
-    await canvasPage.verifyTemplate("Hero's Journey");
-
-    // Colors should be preserved
-    const pinkNote = page.locator('.note:has-text("Pink on standard")');
-    const greenNote = page.locator('.note:has-text("Green on standard")');
-
-    await expect(pinkNote).toHaveClass(/color-pink/);
-    await expect(greenNote).toHaveClass(/color-green/);
-
-    // Switch to another template
-    await canvasPage.switchToTemplate('Now/Next/Future');
-    await canvasPage.verifyTemplate('Now/Next/Future');
-
-    // Colors should still be preserved
-    await expect(pinkNote).toHaveClass(/color-pink/);
-    await expect(greenNote).toHaveClass(/color-green/);
-
-    // Color picker should still work
-    await canvasPage.selectNote(pinkNote);
-    await page.click('.color-swatch[data-color="blue"]');
-    await expect(pinkNote).toHaveClass(/color-blue/);
-  });
-
-  test.skip('Should handle color persistence with large datasets', async ({
-    page,
-  }) => {
-    // Create fewer notes (6 instead of 12) for better CI performance
-    const colors = ['yellow', 'pink', 'green', 'blue'];
-    const noteData = [];
-
-    for (let i = 0; i < 6; i++) {
-      const x = 350 + (i % 3) * 200;
-      const y = 300 + Math.floor(i / 3) * 200;
-      const color = colors[i % 4];
-
-      const note = await canvasPage.createNote(x, y);
-      await canvasPage.selectNote(note);
-      await page.click(`.color-swatch[data-color="${color}"]`);
-      await canvasPage.editNoteContent(`Note ${i + 1}`, note);
-
-      noteData.push({ index: i + 1, color, x, y });
-      await canvasPage.waitForAppReady();
-    }
-
-    // Test persistence by refreshing the page instead of export/import
-    // This tests the same functionality (localStorage persistence) without clipboard issues
-    await page.reload();
-    await page.waitForLoadState('domcontentloaded');
-    await canvasPage.waitForAppReady();
-
-    // Verify all notes and colors are restored after page refresh
-    for (const data of noteData) {
-      const note = page.locator(`.note:has-text("Note ${data.index}")`);
-      await expect(note).toBeVisible();
-      await expect(note).toHaveClass(new RegExp(`color-${data.color}`)); // Fixed CSS pattern
-    }
-  });
-
-  test('Should preserve current color state after operations', async ({
-    page,
-  }) => {
-    // Set initial color
-    await page.click('.color-swatch[data-color="blue"]');
-
-    // Create and edit a note
-    const note = await canvasPage.createNote(400, 300);
-    await canvasPage.editNoteContent('Test note', note);
-
-    // Move the note
-    await canvasPage.selectNote(note);
-    await canvasPage.moveNote(note, 100, 100);
-
-    // Current color should still be blue
-    await expect(page.locator('.color-swatch[data-color="blue"]')).toHaveClass(
-      /active/,
-    );
-
-    // Create another note - should be blue
-    const note2 = await canvasPage.createNote(600, 300);
-    await expect(note2).toHaveClass(/color-blue/);
-  });
-});
