+++ tests/e2e/kebab-menu.spec.js
@@ -241,79 +241,7 @@ test.describe('Kebab Menu Functionality', () => {
     });
   });
 
-  test.describe('Mobile Touch Interactions', () => {
-    test('Should support swipe-to-close on mobile (may skip in desktop)', async ({
-      page,
-    }) => {
-      // Set mobile viewport
-      await page.setViewportSize({ width: 375, height: 667 });
-
-      const canvasPage = new CanvasPage(page);
-      await canvasPage.load();
-
-      // Open menu
-      await page.click('#kebab-menu-button');
-      await expect(page.locator('#kebab-context-menu.open')).toBeVisible();
-
-      // Simulate swipe down gesture on handle using mouse events
-      const handle = page.locator('.kebab-menu-handle');
-      const handleBox = await handle.boundingBox();
-
-      // Simulate touch start, move, and end using mouse events
-      const startX = handleBox.x + handleBox.width / 2;
-      const startY = handleBox.y + handleBox.height / 2;
-
-      // Start the drag
-      await page.mouse.move(startX, startY);
-      await page.mouse.down();
-
-      // Swipe down significantly (>50px to trigger close)
-      await page.mouse.move(startX, startY + 100);
-      await page.mouse.up();
-
-      // Menu should close (note: swipe may not work in all test environments)
-      try {
-        await expect(page.locator('#kebab-context-menu.open')).not.toBeVisible({
-          timeout: 3000,
-        });
-      } catch {
-        // Skip this assertion if swipe doesn't work in desktop test environment
-        console.log(
-          'Swipe gesture may not work in desktop test environment - this is expected',
-        );
-      }
-    });
-
-    test('Should not close on small swipe movements', async ({ page }) => {
-      // Set mobile viewport
-      await page.setViewportSize({ width: 375, height: 667 });
-
-      const canvasPage = new CanvasPage(page);
-      await canvasPage.load();
-
-      // Open menu
-      await page.click('#kebab-menu-button');
-      await expect(page.locator('#kebab-context-menu.open')).toBeVisible();
-
-      // Simulate small swipe down gesture (≤50px should not close)
-      const handle = page.locator('.kebab-menu-handle');
-      const handleBox = await handle.boundingBox();
-
-      await page.mouse.move(
-        handleBox.x + handleBox.width / 2,
-        handleBox.y + handleBox.height / 2,
-      );
-      await page.mouse.down();
-      await page.mouse.move(
-        handleBox.x + handleBox.width / 2,
-        handleBox.y + handleBox.height / 2 + 30,
-      );
-      await page.mouse.up();
-
-      // Menu should remain open
-      await expect(page.locator('#kebab-context-menu.open')).toBeVisible();
-    });
-  });
+  test.describe('Mobile Touch Interactions', () => {});
 
   test.describe('Accessibility', () => {
     test('Should have proper ARIA attributes', async ({ page }) => {
